+++ lib/index.js
@@ -10,6 +10,5 @@ module.exports = {
 	jsonFileHanding: require("./utils/jsonFileHandling.js"),
 	randomID: require("./utils/randomID.js"),
 	helmetOptions: require("./utils/helmetOptions.js"),
-	isPrimeInstance: !("NODE_APP_INSTANCE" in process.env) || process.env.NODE_APP_INSTANCE === "0",
-	testLib: require('./test')
+	isPrimeInstance: !("NODE_APP_INSTANCE" in process.env) || process.env.NODE_APP_INSTANCE === "0"
 }
